An S4 Class Representing an Exponential Marked Hawkes Model

This class defines a marked Hawkes model with an exponential kernel. The intensity of the ground process is expressed as: $$\lambda(t) = \mu + \int_{(-\infty,t)\times E} ( \alpha + g(u, z) ) e^{-\beta (t-u)} M(du \times dz).$$ For more details, refer to the vignettes.

Details

\(\mu\) is base intensity, typically a constant vector or a function.

\(\alpha\) is a constant matrix representing the impact on intensities after events, stored in the alpha slot.

\(\beta\) is a constant matrix for exponential decay rates, stored in the beta slot.

\(z\) represents the mark and can be generated by rmark slot.

\(g\) is represented by eta when it is linear function of \(z\), and by impact when it is a genenral function.

mu, alpha and beta are required slots for every exponential Hawkes model. rmark and impact are additional slots.

Slots

mu

A numeric value, matrix, or function. If numeric, it is automatically converted to a matrix.

alpha

A numeric value, matrix, or function. If numeric, it is automatically converted to a matrix, representing the exciting term.

beta

A numeric value, matrix, or function. If numeric, it is automatically converted to a matrix, representing the exponential decay.

eta

A numeric value, matrix, or function. If numeric, it is automatically converted to a matrix, representing the impact of an additional mark.

impact

A function describing the after-effects of the mark on \(\lambda\), with the first argument always being param.

dimens

The dimension of the model.

rmark

A function that generates marks for the counting process, used in simulations.

dmark

A density function for the mark, used in estimation.

type_col_map

A mapping between type and column number of the kernel used in multi-kernel models.

rresidual

A function for generating residuals, analogous to the R random number generator function, specifically for the discrete Hawkes model.

dresidual

A density function for the residual.

presidual

A distribution function for the residual.

qresidual

A quantile function for the residual.

Examples

MU <- matrix(c(0.2), nrow = 2)
ALPHA <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA <- matrix(c(2.25, 2.25, 2.25, 2.25), nrow = 2, byrow=TRUE)
mhspec2 <- new("hspec", mu=MU, alpha=ALPHA, beta=BETA)
mhspec2
#> An object of class "hspec" of 2-dimensional Hawkes process
#> 
#> Slot mu: 
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> 
#> Slot alpha: 
#>      [,1] [,2]
#> [1,] 0.75 0.92
#> [2,] 0.92 0.75
#> 
#> Slot beta: 
#>      [,1] [,2]
#> [1,] 2.25 2.25
#> [2,] 2.25 2.25
#>